requirements.txt
# Telegram Bot
aiogram==3.4.1
aiohttp==3.9.3

# HTTP Client
httpx==0.27.0

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# PostgreSQL
psycopg2-binary==2.9.9

# Scheduler
apscheduler==3.10.4

# Parsing
beautifulsoup4==4.12.3
lxml==5.1.0

# Environment
python-dotenv==1.0.1

# Logging
colorlog==6.8.2

